feat: cli improvements (#167)
**Problem**
- When running `create-merkle-tree-collection` or
`create-meta-merkle-tree`, the save_path is passed into
read_stake_meta_collection and read_merkle_tree_collection,
respectively, with a full path. This causes the read functions to look
for a file that doesn't exist, as the expected filename gets
concatenated onto the save_path.
- As operators, we would like improved visibility into distribution
account stats for a given epoch

**Solution**
- Pass in an unmodified save_path to read functions
- Add get-tip-distribution-stats for printing a table of distribution
account stats for an epoch